+++ core/execution/future/market.go
@@ -747,7 +747,7 @@ func (m *Market) GetMarketData() types.MarketData {
 	} else {
 		targetStake = m.getTargetStake().String()
 	}
-	bounds := m.pMonitor.GetCurrentBounds()
+	bounds := m.pMonitor.GetBounds()
 	for _, b := range bounds {
 		b.MaxValidPrice = m.priceToMarketPrecision(b.MaxValidPrice) // effictively floors this
 		b.MinValidPrice = m.priceToMarketPrecision(b.MinValidPrice)